requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
supabase>=2.4.0
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.0
anthropic>=0.26.0
apscheduler>=3.10.0
gspread>=6.1.0
google-auth>=2.29.0
pyyaml>=6.0.0
python-multipart>=0.0.9
